We did a lot of things. Answered Q3-13, deleted predict.py and moved it to api.py and modified other files to not break anything,deleted notebooks/,create environment.yml to easily recreate environment, added version requirements to requirements.txt where all requirments are now stored. Added the use of the package mypy as a type checker, and finally made sure that data used is pushed to gcloud and we save the hash in the relevaent run folder in outputs/ such that we can recreate any run
