+++ configs/config.yaml
@@ -5,7 +5,7 @@ training:
   optimizer: adam
   loss_function: cross_entropy
   wandb_api_key : '38cdb79190214b4b26e0a94db6a9f78fa7e3168c'
-  training_image: ${hydra:runtime.cwd}/results
+  training_image: ${hydra:runtime.cwd}/reports/figures
 
 
 model:
