+++ dockerfiles/api.dockerfile
@@ -16,4 +16,4 @@ COPY data/ data/
 
 RUN pip install --default-timeout=300 -r requirements_api.txt --no-cache-dir
 
-ENTRYPOINT ["uvicorn", "src.exam_project.predict:app", "--host", "0.0.0.0", "--port", "8000"]
+ENTRYPOINT ["uvicorn", "src.exam_project.api:app", "--host", "0.0.0.0", "--port", "8000"]
