+++ requirements.txt
@@ -1,25 +1,212 @@
-coverage
-dvc
-dvc-gdrive
-dvc_gs
+absl-py==2.1.0
+aiohappyeyeballs==2.4.4
+aiohttp==3.11.11
+aiohttp-retry==2.9.1
+aiosignal==1.3.2
+amqp==5.3.1
+annotated-types==0.7.0
+antlr4-python3-runtime==4.9.3
+anyio==4.8.0
+appdirs==1.4.4
+asyncssh==2.19.0
+atpublic==5.0
+attrs==24.3.0
+babel==2.16.0
+billiard==4.2.1
+build==1.2.2.post1
+cachetools==5.5.0
+celery==5.4.0
+certifi==2024.12.14
+cffi==1.17.1
+cfgv==3.4.0
+charset-normalizer==3.4.1
+click==8.1.8
+click-didyoumean==0.3.1
+click-plugins==1.1.1
+click-repl==0.3.0
+colorama==0.4.6
+coloredlogs==15.0.1
+configobj==5.0.9
+contourpy==1.3.1
+coverage==7.6.9
+cryptography==43.0.3
+cycler==0.12.1
+decorator==5.1.1
+dictdiffer==0.9.0
+diskcache==5.6.3
+distlib==0.3.9
+distro==1.9.0
+docker-pycreds==0.4.0
+dpath==2.2.0
+dulwich==0.22.7
+dvc==3.59.0
+dvc-data==3.16.8
+dvc-gdrive==3.0.1
+dvc-gs==3.0.1
+dvc-http==2.32.0
+dvc-objects==5.1.0
+dvc-render==1.0.2
+dvc-studio-client==0.21.0
+dvc-task==0.40.2
+entrypoints==0.4
+exam-project==0.0.1
 fastapi==0.115.6
-fastapi
-httpx
-hydra-core
+filelock==3.16.1
+flatbuffers==24.12.23
+flatten-dict==0.4.2
+flufl-lock==8.1.0
+fonttools==4.55.3
+frozenlist==1.5.0
+fsspec==2024.12.0
+funcy==2.0
+gcsfs==2024.12.0
+ghp-import==2.1.0
+gitdb==4.0.12
+gitpython==3.1.44
+google-api-core==2.24.0
+google-api-python-client==2.159.0
+google-auth==2.37.0
+google-auth-httplib2==0.2.0
+google-auth-oauthlib==1.2.1
+google-cloud-core==2.4.1
+google-cloud-storage==2.19.0
+google-crc32c==1.6.0
+google-resumable-media==2.7.2
+googleapis-common-protos==1.66.0
+grandalf==0.8
+griffe==1.5.4
+grpcio==1.69.0
+gto==1.7.2
+h11==0.14.0
+httpcore==1.0.7
+httplib2==0.22.0
+httpx==0.28.1
+huggingface-hub==0.27.1
+humanfriendly==10.0
+hydra-core==1.3.2
+identify==2.6.5
+idna==3.10
+iniconfig==2.0.0
+invoke==2.2.0
+iterative-telemetry==0.0.9
+jinja2==3.1.5
+joblib==1.4.2
 kagglehub==0.3.6
-loguru
+kiwisolver==1.4.8
+kombu==5.4.2
+loguru==0.7.3
+markdown==3.7
+markdown-it-py==3.0.0
+markupsafe==3.0.2
 matplotlib==3.10.0
-numpy<2
-onnx
-onnxruntime
-onnxscript
+mccabe==0.7.0
+mdurl==0.1.2
+mergedeep==1.3.4
+mkdocs==1.5.3
+mkdocs-autorefs==1.3.0
+mkdocs-material==9.4.6
+mkdocs-material-extensions==1.3.1
+mkdocstrings==0.27.0
+mkdocstrings-python==1.12.2
+ml-dtypes==0.5.1
+mpmath==1.3.0
+multidict==6.1.0
+mypy-extensions==1.0.0
+networkx==3.4.2
+nodeenv==1.9.1
+numpy==1.26.4
+oauth2client==4.1.3
+oauthlib==3.2.2
+omegaconf==2.3.0
+onnx==1.17.0
+onnxruntime==1.20.1
+onnxscript==0.1.0.dev20250118
+orjson==3.10.14
+packaging==24.2
+paginate==0.5.7
 pandas==2.2.3
+pathspec==0.12.1
+pillow==11.1.0
+pip-tools==7.4.1
+pipdeptree==2.24.0
+platformdirs==4.3.6
+pluggy==1.5.0
+pre-commit==3.5.0
+prompt-toolkit==3.0.48
+propcache==0.2.1
+proto-plus==1.25.0
+protobuf==5.29.3
+psutil==6.1.1
+pyasn1==0.6.1
+pyasn1-modules==0.4.1
+pycodestyle==2.12.1
+pycparser==2.22
+pydantic==2.10.5
+pydantic-core==2.27.2
+pydot==3.0.4
+pydrive2==1.21.3
+pyflakes==3.2.0
+pygit2==1.17.0
+pygments==2.19.1
+pygtrie==2.5.0
+pymdown-extensions==10.14
+pyopenssl==24.2.1
+pyparsing==3.2.1
+pyproject-hooks==1.2.0
+pytest==7.4.3
+python-dateutil==2.9.0.post0
+pytz==2024.2
+pyyaml==6.0.2
+pyyaml-env-tag==0.1
+regex==2024.11.6
+requests==2.32.3
+requests-oauthlib==2.0.0
+rich==13.9.4
+rsa==4.9
+ruamel-yaml==0.18.10
+ruamel-yaml-clib==0.2.12
+ruff==0.1.3
+safetensors==0.5.2
 scikit-learn==1.6.0
-# Editable
+scipy==1.15.1
+scmrepo==3.3.9
+semver==3.0.2
+sentry-sdk==2.20.0
+setproctitle==1.3.4
+shellingham==1.5.4
+shortuuid==1.0.13
+shtab==1.7.1
+six==1.17.0
+smmap==5.0.2
+sniffio==1.3.1
+sqltrie==0.11.1
+starlette==0.41.3
+svgwrite==1.4.3
+sympy==1.13.3
+tabulate==0.9.0
+tensorboard==2.18.0
+tensorboard-data-server==0.7.2
+threadpoolctl==3.5.0
+tokenizers==0.21.0
+toml==0.10.2
+tomlkit==0.13.2
 torch==2.2.2
-torch_tb_profiler
+torch-tb-profiler==0.4.3
 torchvision==0.17.2
+tqdm==4.67.1
 transformers==4.47.1
 typer==0.15.1
+typing-extensions==4.12.2
+tzdata==2024.2
+uritemplate==4.1.1
+urllib3==2.3.0
 uvicorn==0.34.0
-wandb
+vine==5.1.0
+virtualenv==20.29.0
+voluptuous==0.15.2
+wandb==0.19.3
+watchdog==6.0.0
+wcwidth==0.2.13
+werkzeug==3.1.3
+yarl==1.18.3
+zc-lockfile==3.0.post1
