+++ requirements_dev.txt
@@ -2,13 +2,8 @@
 invoke==2.2.0
 
 
-# docs
-mkdocs==1.5.3
-mkdocs-material==9.4.6
-mkdocstrings-python==1.12.2
 
-# pre-commit
-pre-commit==3.5.0
+
 
 
 # Editable
