+++ tests/test_data.py
@@ -1,7 +1,7 @@
 import numpy as np
-from sklearn.compose import ColumnTransformer
-from exam_project.data import load_and_preprocess_data
-from tests.__init__ import _PATH_DATA  # Adjust this import to your actual utils module
+from sklearn.compose import ColumnTransformer # type: ignore
+from exam_project.data import load_and_preprocess_data # type: ignore
+from tests.__init__ import _PATH_DATA  # type: ignore #Adjust this import to your actual utils module
 
 # new change
